Merged in fix/vfs (pull request #36)
Fix - VFS

Previously, the list of vfs buckets only returned the folder's basename. There was no distinction between a folder and subfolder as they were flattened.

In this new version, we display a relative path to the root directory specified in the uri, including subfolders.
For instance, if we have the following structure:
```
</tmp> # root directory
data/
    |_a.txt
    |_db/
      |_b.txt
```
We get `[data/db, data]` as result of the buckets list.

In the same way, we get the following objects list for the "data" bucket: `[a.txt, db/b.txt]`

Furthermore, we add a test to not upload an object if the targeted bucket doesn't exist.

Approved-by: Cedric Chantepie <c.chantepie@yahoo.fr>
Approved-by: David Colle <david.colle@fabernovel.com>