Merged in fix/s3-query (pull request #45)
S3 - Fix query parameters

Previously, the token was not added to the query string if no maximum limit was set.
Therefore if a bucket contained more than a thousand elements (the default AWS limit), we would request the same elements over and over again.

This commit builds the query string accordingly.

Approved-by: Cedric Chantepie <c.chantepie@yahoo.fr>
Approved-by: David Colle <david.colle@fabernovel.com>